Add files via upload
The problem of interest is the prediction of application rate. Imagine a user visiting Glassdoor, and
performing a job search. From the set of displayed results, the user clicks on certain ones that she is interested in, and after checking job descriptions, she further clicks on apply button therein to land into an application page. The application rate is defined as the fraction of applies (after visiting job description pages), and the goal is to predict this metric using the dataset described in the following section.